refactor(cli): move orchestration infrastructure out of shared into gen-orchestrator
file-scanner, program-factory, and tsconfig-loader had exactly one
consumer (gen-orchestrator/orchestrator.ts) despite living in shared/.
Move them into gen-orchestrator/infra/ so the module layout reflects
actual ownership, and prune the now-dead scanDirectory/ScanOptions/
ScanResult re-exports from shared/index.ts.
