+++ packages/cli/src/gen-orchestrator/infra/file-scanner.ts
@@ -1,7 +1,7 @@
 import { readdir, stat } from "node:fs/promises";
 import { join, matchesGlob, relative, resolve } from "node:path";
-import type { Diagnostic } from "../core/index.js";
-import { toPosixPath } from "./path-utils.js";
+import type { Diagnostic } from "../../core/index.js";
+import { toPosixPath } from "../../shared/path-utils.js";
 
 export interface ScanResult {
   readonly files: ReadonlyArray<string>;
