+++ packages/cli/src/gen-orchestrator/infra/program-factory.ts
@@ -1,5 +1,5 @@
 import ts from "typescript";
-import type { Diagnostic } from "../core/index.js";
+import type { Diagnostic } from "../../core/index.js";
 import { loadTsconfig } from "./tsconfig-loader.js";
 
 const DEFAULT_COMPILER_OPTIONS: ts.CompilerOptions = {
