+++ packages/cli/src/gen-orchestrator/infra/tsconfig-loader.ts
@@ -1,7 +1,7 @@
 import { existsSync } from "node:fs";
 import path from "node:path";
 import ts from "typescript";
-import type { Diagnostic } from "../core/index.js";
+import type { Diagnostic } from "../../core/index.js";
 
 export interface LoadTsconfigOptions {
   readonly cwd: string;
